Update source as well as documentation with respect to slurm 20.02.4 source
With some minor modifications it is possible to patch the slurm sources to build rpms through rpmbuild

The patched version was tested with a containerized centos7 slurm 20.02.4 and a single A100

It was possible to start multiple single as well as multi Aurora jobs simultaneously on the A100.
ve_exec in combination with the ''VE_NODE_NUMBER'' environment variable were used for this.

The usage of the Auroras were successfully restricted to the assigned ones.

MISSING TESTS:
- multinode
- MPI